Migrated the createTestImage and randomPixelLocations from test_lon_lat.py to unittesting_tools.py.
